requirements.txt
fastapi==0.104.0
uvicorn[standard]==0.23.2
redis==6.4.0
sqlalchemy==2.0.22
pydantic==2.4.2
python-dotenv==1.0.0
aiohttp==3.8.5
aiosqlite==0.19.0
numpy==1.24.3
pandas==2.1.1
openpyxl==3.1.2
APScheduler==3.10.4
pytz==2023.3.post1
python-multipart==0.0.6
asyncio==3.4.3
backoff==2.2.1
yfinance==0.2.31
requests==2.31.0
alpha_vantage==2.3.1
